requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-script
flask-marshmallow
flask-bcrypt 
pyjwt 
psycopg2-binary
flask-restful 
flask-jwt-extended 
dnspython 
uwsgi
flask_cors
flask-script
flask-migrate
python-dotenv
marshmallow-sqlalchemy
requests
flask-socketio
celery
# redis
